+++ src/main/java/janggi/domain/board/Dynasty.java
@@ -1,4 +1,4 @@
-package janggi.domain;
+package janggi.domain.board;
 
 public enum Dynasty {
     CHU,
