+++ src/main/java/janggi/domain/piece/Chariot.java
@@ -1,10 +1,10 @@
 package janggi.domain.piece;
 
-import janggi.domain.Dynasty;
+import janggi.domain.board.Dynasty;
 import janggi.domain.board.Direction;
-import janggi.domain.board.Point;
 import janggi.domain.piece.move.MoveStrategy;
 import janggi.domain.piece.move.OrMoveStrategy;
+import janggi.domain.piece.move.Path;
 import janggi.domain.piece.move.strategy.NoObstacleStrategy;
 import janggi.domain.piece.move.strategy.PalaceAreaStrategy;
 import java.util.List;
