+++ src/main/java/janggi/domain/piece/move/PathCalculator.java
@@ -1,7 +1,6 @@
 package janggi.domain.piece.move;
 
 import janggi.domain.board.Point;
-import janggi.domain.piece.Path;
 
 @FunctionalInterface
 public interface PathCalculator {
