+++ src/main/java/janggi/domain/player/Player.java
@@ -1,6 +1,6 @@
 package janggi.domain.player;
 
-import janggi.domain.Dynasty;
+import janggi.domain.board.Dynasty;
 import java.util.Objects;
 
 public class Player {
