+++ src/main/java/janggi/view/InitializeView.java
@@ -1,6 +1,6 @@
 package janggi.view;
 
-import janggi.domain.Dynasty;
+import janggi.domain.board.Dynasty;
 import janggi.domain.board.BoardSetUp;
 import janggi.domain.player.Player;
 import java.util.Scanner;
