+++ src/test/java/janggi/domain/board/BoardSetUpTest.java
@@ -2,7 +2,6 @@ package janggi.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.domain.Dynasty;
 import janggi.domain.piece.Elephant;
 import janggi.domain.piece.Horse;
 import janggi.domain.piece.Piece;
