+++ src/test/java/janggi/domain/player/PlayerTest.java
@@ -3,7 +3,7 @@ package janggi.domain.player;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import janggi.domain.Dynasty;
+import janggi.domain.board.Dynasty;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
